Increased robustness of lobby manager
The lobby manager now checks to see if it is already in process of
registering to a tournament.

The snapshot processing function will now stop everything if something
goes wrong. This is implemented through a global stop flag using the boost
IPC stuff. All the old asserts are now handled in release mode as well.

After registration count changes, the old and new values are logged.


Former-commit-id: 8fdd6025173f6bdedcc9e6c5e064aca80b1bda81