Improved popup window handling
Moved popup window stuff to window_utils and changed the way the closing
is performed to try pressing a close button within the dialog first. If
none is found or the window is still open, a WM_CLOSE is sent.

This also fixes a problem where when a registration would fail, the
program would get stuck due to not understanding the failure.


Former-commit-id: 431fd9a569dbb44100242f35dc49cdcbeeb01ad2