Add support for small variations in text positions
This changes how text and cards are read from a screen capture. The values
in the XML setting file are the topmost coordinates from where a capture
is attempted. The rank font is now also compared against a monochrome
image. Suit colors are searched for from within the whole card area.


Former-commit-id: 5c1d99eaa2c99b386e72f0bae4b092c27897ab45