Fix scheduler issues due to randomization
There was a problem where the search for the next span didn't break at the
first (earliest) match. Another issue comes from the fact that the time to
next activity/break is randomized, which could lead to multiple
activations of a single activity span if the randomized time was less than
the time to the next span.


Former-commit-id: dae16aec8c18b274e104f49b1a2fe6076ec4a3bc