Fix incorrect handling of table replacements
The problem occurred when a previous table was closed and a new table
opened fast enought to replace it. In this case the table count remained
the same, but the table id was different. This lead to the registration
count being decreased correctly but not increased to accommodate the new
table.


Former-commit-id: a5c45fdbb567df143b12a6c8f57bf6a2e67d8249