Fix detection of delayed popup closing
In some cases the registration popup was closed with a delay, which lead
to the registration being wrongfully stuck in a popup waiting state.

The states traversed in this case were:
REGISTER -> CLOSE_POPUP -> WAIT_POPUP -> CLOSE_POPUP -> timeout

Now the states go as follows:
REGISTER -> CHECK_POPUP -> CLOSE_POPUP -> WAIT_POPUP -> ...

Where CHECK_POPUP does the actual waiting for the popup to appear, and
does the timeout checks as well.


Former-commit-id: 5a9ab6e5d01349ea2a8846c4ce5ad08ae87d4045